requirements.txt
Django==3.1.1
djangorestframework==3.11.1
django-environ
apscheduler==3.6.3
psycopg2-binary==2.8.5
gunicorn==19.9.0
dj-database-url==0.5.0
whitenoise==4.1.2